refactor: reorganize file structure according to data, domain, presentation convensions
