+++ lib/app/app.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:worldreader/core/app_theme/app_theme.dart';
-import 'package:worldreader/view/dashboard.dart';
-import 'package:worldreader/view/login_page.dart';
-import 'package:worldreader/view/on_boarding_screen.dart';
-import 'package:worldreader/view/sign_up_page.dart';
+import 'package:worldreader/features/auth/presentation/view/login_page.dart';
+import 'package:worldreader/features/auth/presentation/view/sign_up_page.dart';
+import 'package:worldreader/features/home/presentation/view/dashboard.dart';
+import 'package:worldreader/features/splash/presentation/view/on_boarding_screen.dart';
 
 class MyApp extends StatelessWidget {
   const MyApp({super.key});
