+++ lib/main.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
-import 'app.dart';
+
+import 'app/app.dart';
 
 void main() {
   runApp(
